重构 Electron 桌宠入口模块
