+++ tests/electron-main-dashboard-url.test.ts
@@ -4,6 +4,8 @@ import { join } from 'node:path';
 import test from 'node:test';
 
 const electronMain = readFileSync(join(process.cwd(), 'electron/main.js'), 'utf-8');
+const applicationMenu = readFileSync(join(process.cwd(), 'electron/application-menu.js'), 'utf-8');
+const trayModule = readFileSync(join(process.cwd(), 'electron/tray.js'), 'utf-8');
 
 test('electron opens the local dashboard through stable IPv4 loopback', () => {
   assert.match(electronMain, /mainWindow\.loadURL\(`http:\/\/127\.0\.0\.1:\$\{DASHBOARD_PORT\}`\)/);
@@ -11,26 +13,26 @@ test('electron opens the local dashboard through stable IPv4 loopback', () => {
 });
 
 test('electron uses Chinese menus with close-to-tray control', () => {
-  assert.match(electronMain, /function createApplicationMenu\(\)/);
-  assert.match(electronMain, /Menu\.setApplicationMenu\(Menu\.buildFromTemplate\(template\)\)/);
-  assert.match(electronMain, /label: '文件'/);
-  assert.match(electronMain, /label: '编辑'/);
-  assert.match(electronMain, /label: '视图'/);
-  assert.match(electronMain, /label: '窗口'/);
-  assert.match(electronMain, /label: '帮助'/);
-  assert.match(electronMain, /label: '点 × 后隐藏到后台'/);
-  assert.match(electronMain, /type: 'checkbox'/);
-  assert.match(electronMain, /writeCloseToTrayPreference\(menuItem\.checked\)/);
-  assert.doesNotMatch(electronMain, /label: 'File'/);
-  assert.doesNotMatch(electronMain, /label: 'Edit'/);
-  assert.doesNotMatch(electronMain, /label: 'View'/);
-  assert.doesNotMatch(electronMain, /label: 'Window'/);
-  assert.doesNotMatch(electronMain, /label: 'Help'/);
+  assert.match(applicationMenu, /function createApplicationMenu/);
+  assert.match(applicationMenu, /Menu\.setApplicationMenu\(Menu\.buildFromTemplate\(template\)\)/);
+  assert.match(applicationMenu, /label: '文件'/);
+  assert.match(applicationMenu, /label: '编辑'/);
+  assert.match(applicationMenu, /label: '视图'/);
+  assert.match(applicationMenu, /label: '窗口'/);
+  assert.match(applicationMenu, /label: '帮助'/);
+  assert.match(applicationMenu, /label: '点 × 后隐藏到后台'/);
+  assert.match(applicationMenu, /type: 'checkbox'/);
+  assert.match(applicationMenu, /writeCloseToTrayPreference\(menuItem\.checked\)/);
+  assert.doesNotMatch(applicationMenu, /label: 'File'/);
+  assert.doesNotMatch(applicationMenu, /label: 'Edit'/);
+  assert.doesNotMatch(applicationMenu, /label: 'View'/);
+  assert.doesNotMatch(applicationMenu, /label: 'Window'/);
+  assert.doesNotMatch(applicationMenu, /label: 'Help'/);
 });
 
 test('electron changes cwd to userData before reading close-to-tray menu preference', () => {
   const chdirIndex = electronMain.indexOf('process.chdir(userDataPath)');
-  const menuCallIndex = electronMain.indexOf('createApplicationMenu();');
+  const menuCallIndex = electronMain.indexOf('createApplicationMenu({');
 
   assert.notEqual(chdirIndex, -1);
   assert.notEqual(menuCallIndex, -1);
@@ -41,7 +43,7 @@ test('electron changes cwd to userData before reading close-to-tray menu prefere
 test('electron tray uses app icons and notifies after background hide', () => {
   assert.match(electronMain, /function createTrayIcon\(\)/);
   assert.match(electronMain, /build-resources\/icon\.ico/);
-  assert.match(electronMain, /new Tray\(createTrayIcon\(\)\)/);
+  assert.match(trayModule, /new Tray\(createTrayIcon\(\)\)/);
   assert.match(electronMain, /function notifyWindowHidden\(\)/);
   assert.match(electronMain, /tray\.displayBalloon/);
   assert.match(electronMain, /CatsCo 已在后台运行/);
